# backend-go/api/openapi.yaml
openapi: 3.0.3
info:
  title: xiaoan-platform API
  version: 2.0.0
  description: |
    xiaoan-platform backend contract (Go control plane).

    This file is the single source of truth for the HTTP API.
    Go server types and TypeScript clients are generated from it.

    ## Error envelope conventions (legacy-compatible)

    The React frontend was validated against the Django reference
    implementation, which produced four distinct error shapes:

    - `{"detail": "message"}` — 403/404 style errors (DRF PermissionDenied / NotFound).
    - `{"error": "message"}` — simple JSON errors (identity endpoints, 409/502 conflicts).
    - `{"field": ["message", ...]}` — field validation errors (DRF serializer style).
    - `["message"]` — bare array, 400 for hand-raised DRF ValidationErrors (e.g. "application_id is required").

    All four shapes are reproduced exactly. Clients should read the
    `detail` / `error` / first field message in that order.

    ## Authentication

    Normal users authenticate with the HttpOnly opaque session cookie
    `studio_session` (set by OAuth exchange / admin login). All
    state-changing requests additionally require the CSRF header
    `X-CSRF-Token` matching the `studio_csrf` cookie (double-submit).
    `GET /api/v2/artifacts/{id}/open` is cookie-authenticated because it is
    opened via browser navigation (window.open) and cannot send headers.

    ## SSE

    `GET /api/v2/runs/{id}/stream` responds with `text/event-stream`.
...
servers:
...
paths:
...
  /api/identity/oauth/start:
    get:
      operationId: oauthStart
      tags: [identity]
      summary: Redirect (302) the browser to the Feishu OAuth authorize page
      parameters:
        - name: return_to
          in: query
          schema: { type: string, default: "/" }
          description: SPA path to return to after login; must start with "/" and not "//".
      responses:
        "302":
          description: Redirect to Feishu authorize URL (state signed, 10 min TTL).
          headers:
            Location:
              schema: { type: string }
        "500":
          description: OAuth not configured
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SimpleError" }
  /api/identity/oauth/exchange:
    get:
      operationId: oauthExchange
      tags: [identity]
      summary: Exchange the OAuth ?code&state for a studio session
      description: |
        Verifies the signed state, exchanges the code at the Feishu token
        endpoint, upserts the local user (auth_source=feishu), stores the
        refresh token AES-256-GCM encrypted, sets the `studio_session`
        HttpOnly cookie and returns the session user plus JWT tokens for the
        legacy transition window.
      parameters:
        - name: code
          in: query
          required: true
          schema: { type: string }
        - name: state
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Authenticated
          headers:
            Set-Cookie:
              schema: { type: string }
          content:
            application/json:
              schema:
                type: object
                required: [user, tokens, return_to]
                properties:
                  user: { $ref: "#/components/schemas/User" }
                  tokens: { $ref: "#/components/schemas/Tokens" }
                  return_to: { type: string }
        "400":
          description: missing code / invalid or expired oauth state
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SimpleError" }
        "502":
          description: oauth exchange failed / no access token / user info failed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SimpleError" }
...
  /api/identity/admin/login:
    post:
      operationId: adminLogin
      tags: [identity]
      summary: Local admin login (Argon2id password)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string, format: password }
      responses:
        "200":
          description: Admin session established
          content:
            application/json:
              schema:
                type: object
                required: [id, username, is_staff, auth_source]
                properties:
                  id: { type: integer }
                  username: { type: string }
                  is_staff: { type: boolean }
                  auth_source: { type: string }
                  display_name: { type: string }
                  display_id: { type: string }
        "401":
          description: invalid credentials
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SimpleError" }

  # ─────────────────────────────────────────────────────────────── auth ──
  /api/auth/login/:
    post:
      operationId: authLogin
      tags: [auth]
      summary: Local username/password login (transition window)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string, format: password }
      responses:
        "200":
          description: Logged in; also sets studio_session cookie
          content:
            application/json:
              schema:
                type: object
                required: [user, tokens]
                properties:
                  user: { $ref: "#/components/schemas/User" }
                  tokens: { $ref: "#/components/schemas/Tokens" }
        "401":
          description: bad credentials
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetailError" }
...
  /api/auth/token/refresh/:
    post:
      operationId: tokenRefresh
      tags: [auth]
      summary: Rotate a JWT refresh token (transition window)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh]
              properties:
                refresh: { type: string }
      responses:
        "200":
          description: New access token
          content:
            application/json:
              schema:
                type: object
                properties:
                  access: { type: string }
                  refresh: { type: string }
...
  /api/v2/applications:
    get:
      operationId: listApplications
      tags: [applications]
      summary: List openable applications (bare array, no pagination envelope)
      parameters:
        - name: kind
          in: query
          schema: { type: string, enum: [chat, task, custom, all], default: chat }
          description: Invalid values fall back to chat.
        - name: scope
          in: query
          schema: { type: string, enum: [accessible, public, mine, manage], default: public }
          description: Invalid values fall back to public.
        - name: include_unbound
          in: query
          schema: { type: string }
          description: Truthy when case-insensitively in {1,true,yes}.
      responses:
        "200":
          description: Application list items (bare array)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ApplicationListItem" }
    post:
      operationId: createApplication
      tags: [applications]
      summary: Create an application (admin only — agent marketplace "new Aily agent")
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ApplicationCreateRequest" }
      responses:
        "201":
          description: Created (compact detail shape)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApplicationDetail" }
        "400":
          description: field errors
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FieldErrors" }
        "403":
          description: only admins may add agents
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetailError" }
  /api/v2/applications/page:
    get:
      operationId: listApplicationPage
      tags: [applications]
      summary: Keyset-paginated application catalog page (执行报告 §14–§17)
      description: |
        True server-side pagination: LIMIT / cursor-condition evaluate inside
        SQL, one anti-joined query per page (no N+1), and personal usage /
        favorites aggregate ONLY the ids on the returned page. The legacy
        GET /api/v2/applications keeps its whole-catalog array shape for
        callers not yet migrated; new surfaces must use this endpoint.
      parameters:
        - name: kind
          in: query
          schema: { type: string, enum: [chat, fixed, all], default: chat }
          description: >
            `fixed` means kind <> 'chat' (应用中心: pages/forms/dashboards/tasks).
            Invalid values fall back to chat.
        - name: scope
          in: query
          schema: { type: string, enum: [accessible, public, mine, manage], default: public }
          description: Invalid values fall back to public.
        - name: mode
          in: query
          schema: { type: string, enum: [manage, consume], default: manage }
          description: >
            WHO may see the row is `scope`; whether the row may actually be
            USED is `mode` (二次复审 P0-5). `manage` (default) is the
            management surface (智能体市场 / 应用中心): staff also sees
            disabled, private and binding-less rows so they can repair them.
            `consume` is every surface that opens or runs something
            (切换器 / 移动端目录 / bootstrap 分组): it requires
            `enabled = 1` and, for `kind = chat`, an enabled runtime binding
            whose provider exists and is active — exactly the predicate
            `AuthorizeExecution` enforces, so the UI can no longer offer an
            application that the run API then refuses.
        - name: include_unbound
          in: query
          schema: { type: string }
          description: >
            Truthy when case-insensitively in {1,true,yes}; keeps binding-less
            CHAT rows (the management chat catalog's knob). It no longer
            affects `fixed` rows in any mode: a fixed application never needs
            a binding to be consumable, so `kind=fixed&mode=consume` returns
            unbound fixed applications whether or not this flag is set
            (三次复审 P1-R4).
        - name: q
          in: query
          schema: { type: string }
          description: Case-insensitive substring match on name / description / category name.
        - name: category_slug
          in: query
          schema: { type: string }
          description: >
            Exact category match; the sentinel `__uncategorized__` matches rows
            without any category (mobile 其他 tab).
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 24 }
        - name: cursor
          in: query
          schema: { type: string }
          description: Opaque next_cursor from a previous response (base64 JSON of created_at+id).
      responses:
        "200":
          description: One page of application list items (same item shape as the legacy array)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApplicationPage" }
        "400":
          description: invalid cursor
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetailError" }
  /api/v2/applications/resolve:
    get:
      operationId: resolveApplication
      tags: [applications]
      summary: Resolve ONE application by slug or id (执行报告 §12)
      description: >
        Exact single-row resolution for deep links (/chat/:slug, /app/:slug) and
        for the application entity cache. Visibility is evaluated with the
        enterprise access policy (legacy is_public while the rollout flag is off;
        Directory + Application ACL when enabled; staff bypass ACL); then consumption eligibility is applied: the application
        must be enabled and a chat application must have an enabled binding
        backed by an active provider. It answers 404 when the application does
        not exist OR is not visible/consumable — existence is never leaked. This is what makes
        "open /chat/sales" cost one small request instead of a catalog
        download plus Array.find().
      parameters:
        - name: slug
          in: query
          schema: { type: string }
          description: Exact slug. Provide this or `id`.
        - name: id
          in: query
          schema: { type: integer }
          description: Exact application id. Provide this or `slug`.
      responses:
        "200":
          description: The application, in the same shape the catalog page uses
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApplicationListItem" }
        "400":
          description: neither slug nor id was provided
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetailError" }
        "401":
          description: not authenticated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetailError" }
        "404":
          description: unknown, not visible, or not usable by this caller
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetailError" }
  /api/v2/applications/resolve-mention:
    get:
      operationId: resolveApplicationMention
      tags: [applications]
      summary: Resolve the `@` token of a composer instruction to candidates (§14)
      description: >
        Server-side replacement for the client-side @mention candidate array
        (which was built by downloading every application). Matching is
        case-insensitive on the application name and slug; the response is
        ranked (exact name/slug first, then name prefix, then slug prefix,
        then substring) and capped, so the composer never needs the whole catalog
        to route `@销售助手 帮我查一下`.

        Like `resolve`, this is a CONSUMPTION entry point (二次复审 P0-5): only
        usable applications are candidates, so `@` can never switch the
        composer to an application the next send would be refused for.
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string }
          description: The text after `@`, e.g. `销售助手`.
      responses:
        "200":
          description: Ranked candidates (at most 10)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/MentionCandidate" }
        "400":
          description: q is required
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetailError" }
  /api/v2/applications/{id}:
    parameters:
      - $ref: "#/components/parameters/ApplicationId"
    get:
      operationId: getApplication
      tags: [applications]
      summary: Application detail (compact authoring shape, manage-only)
      description: >
        Carries provider-authoring fields (`external_resource_id`,
        `identity_mode`, `execution_mode`), so it is NOT a public read: the
        caller must be authenticated and must be able to MANAGE the
        application (staff only). Visible-to-all
        used to hand every logged-in caller back the authoring fields the
        consumer DTO split had removed from list/resolve. Consumers read
        through /applications/resolve, /applications/page and
        /workspace/bootstrap instead. Non-manageable and nonexistent both
        answer 404 — otherwise the sequential id would let any logged-in
        user enumerate every agent's provider resource id.
      responses:
        "200":
          description: Detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApplicationDetail" }
        "401":
          description: not authenticated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetailError" }
        "404":
          description: unknown or not manageable by this caller
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetailError" }
    patch:
      operationId: updateApplication
      tags: [applications]
      summary: Update an application / rebind runtime (staff only)
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ApplicationUpdateRequest" }
      responses:
        "200":
          description: Updated (compact shape)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApplicationDetail" }
        "404":
          description: unknown or not manageable (opaque — sequential ids must not be enumerable, 三次复审 P0-R4.2)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetailError" }
    delete:
      operationId: deleteApplication
      tags: [applications]
      summary: Delete an application (409 when referenced by conversations/runs)
      responses:
        "204":
          description: Deleted
        "404":
          description: unknown or not manageable (opaque — 三次复审 P0-R4.2)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetailError" }
        "409":
          description: has conversations or runs
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetailError" }
  /api/v2/applications/{id}/favorite:
    parameters:
      - $ref: "#/components/parameters/ApplicationId"
    post:
      operationId: favoriteApplication
      tags: [applications]
      summary: Pin application (idempotent, visibility-checked)
      description: >
        三次复审 P0-R4.3: the favorite relation exists only for applications
        the caller can SEE, so a hidden-but-real id answers the same opaque
        404 as an unknown one — favoriting must not become an existence
        oracle. Fixed applications keep their favorite relation (the ✩).
      responses:
        "200":
          description: Favorite state
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FavoriteResult" }
        "404":
          description: application not found (or not visible — indistinguishable on purpose)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetailError" }
    delete:
      operationId: unfavoriteApplication
      tags: [applications]
      summary: Unpin application (idempotent, visibility-checked)
      responses:
        "200":
          description: Favorite state
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FavoriteResult" }
        "404":
          description: application not found (or not visible — indistinguishable on purpose)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetailError" }
...
  /api/v2/applications/{id}/default-agent:
    parameters:
      - $ref: "#/components/parameters/ApplicationId"
    post:
      operationId: setDefaultAgent
      tags: [applications]
      summary: Set as workspace main agent (chat + enabled binding only)
      responses:
        "200":
          description: Updated (compact shape)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApplicationDetail" }
        "400":
          description: not eligible
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetailError" }
        "403":
          description: staff only
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetailError" }
    delete:
      operationId: unsetDefaultAgent
      tags: [applications]
      summary: Unset workspace main agent
      responses:
        "200":
          description: Updated (compact shape)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApplicationDetail" }
  /api/v2/applications/{id}/attachments:
    parameters:
      - $ref: "#/components/parameters/ApplicationId"
    post:
      operationId: uploadApplicationAttachment
      tags: [applications]
      summary: Pre-run attachment upload (stored in studio storage; the worker uploads to the provider with the user's UAT)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                type:
                  type: string
                  enum: [image, file, feishu_doc, bitable]
                  default: file
                doc_url:
                  type: string
      responses:
        "201":
          description: Attachment created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RuntimeAttachment" }
        "400":
          description: file or doc_url required / binding missing
          content:
            application/json:
              schema:
                oneOf:
                  - { $ref: "#/components/schemas/DetailError" }
                  - { $ref: "#/components/schemas/BareError" }
                  - { $ref: "#/components/schemas/FieldErrors" }
        "502":
          description: attachment upload failed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SimpleError" }

  # ───────────────────────────────────────────────────────── schedules ──
...
  /api/v2/schedules/{id}/enable:
    parameters:
      - $ref: "#/components/parameters/ScheduleId"
    post:
      operationId: enableSchedule
      tags: [schedules]
      summary: Enable a schedule (next_run_at recomputed from now; missed slots are not replayed)
      responses:
        "200":
          description: Updated schedule
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Schedule" }
        "404":
          description: unknown or foreign schedule
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetailError" }
...
components:
...
  schemas:
...
    DetailError:
      type: object
      required: [detail]
      properties:
        detail: { type: string }
    SimpleError:
      type: object
      required: [error]
      properties:
        error: { type: string }
    BareError:
      type: array
      items: { type: string }
...
    Tokens:
      type: object
      properties:
        access: { type: string }
        refresh: { type: string }
...